CI修正: add_joint_methodsテストの不要なキャストを除去
clippy::unnecessary_cast(-D warnings)。`chassis`は`spawn_box`の戻り値
(usize)なので`chassis as usize`は冗長——ローカルのclippyキャッシュが
このテスト追加後の差分を拾わずクリーンだと誤判定していた(CIで検出)。

+++ crates/sim-wasm/src/lib.rs
@@ -2944,17 +2944,7 @@ mod tests {
             .expect("slider joint must succeed");
         world
             .add_wheel_joint(
-                chassis as usize,
-                wheel,
-                1.0,
-                0.0,
-                1.0,
-                0.4,
-                2.5,
-                0.7,
-                0.0,
-                12.0,
-                200.0,
+                chassis, wheel, 1.0, 0.0, 1.0, 0.4, 2.5, 0.7, 0.0, 12.0, 200.0,
             )
             .expect("wheel joint must succeed");
         world
